refactor(persons): EIN gemeinsames PersonSettings-Menue (Tabelle + Detail identisch) (#48)
Die zwei Bearbeiten-Menues waren noch unterschiedlich (Tabelle: nur Felder,
Titel 'Person bearbeiten'; Detail: Felder + Urlaubskontingente, Titel
'Personeneinstellungen'). Jetzt genau EIN Objekt:

- Neue Komponente PersonSettings (Modal 'Personeneinstellungen'): Personenfelder +
  Urlaubskontingente + kompakte, editierbare Projektzuweisungen-Liste. Self-contained
  (eigene Queries/Mutations/Invalidierung), in Tabelle UND Detailansicht identisch.
- Projektzuweisung damit ueber ZWEI Pfade editierbar: die Projekte-Registerkarte
  (unveraendert) und das Einstellungs-Menue (kompakte Liste, aehnlich Urlaubskontingente).
- Sub-Formulare (ContingentForm, MembershipForm, MembershipEditForm) nach
  components/person/forms.tsx ausgelagert und von beiden Stellen genutzt (keine Dopplung).
- PersonDetailPage: lokale Formulare/Kontingent-Logik entfernt, delegiert ans gemeinsame
  Menue; Registerkarten Abwesenheiten/Projekte bleiben wie sie sind.
